refactor: simplify user followers, following, and payments components by integrating useTableSorting hook
