+++ src/pages_layer/user-profile/user-followers/ui/UserFollowers.tsx
@@ -3,82 +3,24 @@
 import type { FollowersListTableHeadKeysType } from '../model/table-headers.types'
 
 import { type Get_FollowersQuery, useGet_FollowersQuery } from '@/src/entities/user'
-import { SortDirection } from '@/src/shared/api'
-import { Link, PROFILE, usePathname, useRouter } from '@/src/shared/routes'
+import { Link, PROFILE } from '@/src/shared/routes'
 import { isGraphQLError } from '@/src/shared/types'
 import {
   DataTable,
-  PAGE_PARAM_KEY,
-  PAGE_SIZE_PARAM_KEY,
-  SORT_BY_PARAM_KEY,
-  SORT_PARAM_KEY,
-  SortDirectionTable,
   type TableColumn,
   paginationPageSize,
+  useTableSorting,
 } from '@/src/widgets/table'
 import { ProfileTabValues } from '@/src/widgets/tabs'
 import { Button } from '@meetgram/ui-kit'
 import { dateFormatting } from '@meetgram/utils'
-import { useParams, useSearchParams } from 'next/navigation'
+import { useParams } from 'next/navigation'
 import { useLocale } from 'next-intl'
 
 export const UserFollowers = () => {
   const locale = useLocale()
-  const searchParams = useSearchParams()
-  const pathname = usePathname()
-  const router = useRouter()
-
-  const itemsPerPage = searchParams.get(PAGE_SIZE_PARAM_KEY) || paginationPageSize[1]
-  const sortDirParam = searchParams.get(SORT_PARAM_KEY)
-  const sortDir = sortDirParam ? +sortDirParam : SortDirectionTable.DESC
-  const sortBy = searchParams.get(SORT_BY_PARAM_KEY) || 'createdAt'
-  const page = searchParams.get(PAGE_PARAM_KEY) || 1
-  const params = new URLSearchParams(searchParams)
-
   const paramsId = useParams()
   const parsedUserId = paramsId?.id ? Number(paramsId.id) : null
-
-  const { data, error, loading } = useGet_FollowersQuery({
-    variables: {
-      pageNumber: +page,
-      pageSize: +itemsPerPage,
-      sortBy,
-      sortDirection: sortDir === SortDirectionTable.DESC ? SortDirection.Desc : SortDirection.Asc,
-      userId: parsedUserId!,
-    },
-  })
-
-  const saveSearchParams = () => {
-    router.replace(`${pathname}?${params.toString()}`)
-  }
-
-  const handleOnPageChange = (page: number) => {
-    params.set(PAGE_PARAM_KEY, String(page))
-
-    saveSearchParams()
-  }
-
-  const handleItemsPerPageChange = (itemsPerPage: number) => {
-    params.set(PAGE_SIZE_PARAM_KEY, String(itemsPerPage))
-
-    saveSearchParams()
-  }
-
-  const handleChangeSorting = (column: string) => {
-    const sortDirection =
-      // eslint-disable-next-line no-nested-ternary
-      sortBy === column
-        ? sortDir === SortDirectionTable.DESC
-          ? SortDirectionTable.ASC
-          : SortDirectionTable.DESC
-        : SortDirectionTable.ASC
-
-    params.set(SORT_PARAM_KEY, String(sortDirection))
-    params.set(SORT_BY_PARAM_KEY, column)
-
-    saveSearchParams()
-  }
-
   const followersListTableHeaders: TableColumn<
     Get_FollowersQuery['getFollowers']['items'][0],
     FollowersListTableHeadKeysType
@@ -118,6 +60,27 @@ export const UserFollowers = () => {
       render: following => dateFormatting(following.createdAt, { locale }),
     },
   ]
+
+  const {
+    currentPage,
+    handleChangeSorting,
+    handleItemsPerPageChange,
+    handleOnPageChange,
+    itemsPerPage,
+    sortBy,
+    sortDir,
+    sortDirection,
+  } = useTableSorting({ defaultSortBy: followersListTableHeaders[3].key })
+
+  const { data, error, loading } = useGet_FollowersQuery({
+    variables: {
+      pageNumber: currentPage,
+      pageSize: itemsPerPage,
+      sortBy,
+      sortDirection,
+      userId: parsedUserId!,
+    },
+  })
 
   return (
     <DataTable
