+++ src/widgets/table/index.ts
@@ -1,3 +1,8 @@
 export * from './const/pagination-config'
 export { type TableActionHeader } from './const/table-action-header'
+export {
+  type UseTableSortingProps,
+  type UseTableSortingResult,
+  useTableSorting,
+} from './lib/useTableSorting'
 export { DataTable, SortDirectionTable, type TableColumn } from './ui/DataTable'
